+++ src/backend/mod.rs
@@ -1,3 +1,3 @@
-pub mod backend;
+pub mod data;
 pub mod geo_from_str;
 pub mod interval;
